removing TIME steps without all baselines
There is an error somewhere in out the miriad visibilities have been
formed. Perhaps from the same visbility file being concatenated twice?

This would cause the mstransform stage to fall over with a core dump.
There was nothing emitted in the casa logger either, which made it
particularly annoying to try to track down.

In anycase, this is a simple check to ensure that there is a conformant
MS.